move get_callbacks
